Move PluggableAction to trigger helpers
