+++ homeassistant/components/philips_js/__init__.py
@@ -2,10 +2,9 @@
 from __future__ import annotations
 
 import asyncio
-from collections.abc import Callable, Coroutine, Mapping
+from collections.abc import Mapping
 from datetime import timedelta
 import logging
-from typing import Any
 
 from haphilipsjs import AutenticationFailure, ConnectionFailure, PhilipsTV
 from haphilipsjs.typing import SystemType
@@ -18,9 +17,9 @@ from homeassistant.const import (
     CONF_USERNAME,
     Platform,
 )
-from homeassistant.core import Context, HassJob, HomeAssistant, callback
+from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.debounce import Debouncer
-from homeassistant.helpers.trigger import TriggerActionType
+from homeassistant.helpers.trigger import PluggableAction
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
 from .const import CONF_ALLOW_NOTIFY, CONF_SYSTEM, DOMAIN
@@ -76,42 +75,6 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         hass.data[DOMAIN].pop(entry.entry_id)
 
     return unload_ok
-
-
-class PluggableAction:
-    """A pluggable action handler."""
-
-    def __init__(self, update: Callable[[], None]) -> None:
-        """Initialize."""
-        self._update = update
-        self._actions: dict[
-            Any, tuple[HassJob[..., Coroutine[Any, Any, None]], dict[str, Any]]
-        ] = {}
-
-    def __bool__(self):
-        """Return if we have something attached."""
-        return bool(self._actions)
-
-    @callback
-    def async_attach(self, action: TriggerActionType, variables: dict[str, Any]):
-        """Attach a device trigger for turn on."""
-
-        @callback
-        def _remove():
-            del self._actions[_remove]
-            self._update()
-
-        job = HassJob(action)
-
-        self._actions[_remove] = (job, variables)
-        self._update()
-
-        return _remove
-
-    async def async_run(self, hass: HomeAssistant, context: Context | None = None):
-        """Run all turn on triggers."""
-        for job, variables in self._actions.values():
-            hass.async_run_hass_job(job, variables, context)
 
 
 class PhilipsTVDataUpdateCoordinator(DataUpdateCoordinator[None]):
