+++ homeassistant/components/philips_js/media_player.py
@@ -152,7 +152,7 @@ class PhilipsTVMediaPlayer(
             await self._tv.setPowerState("On")
             self._state = MediaPlayerState.ON
         else:
-            await self.coordinator.turn_on.async_run(self.hass, self._context)
+            self.coordinator.turn_on.async_run(self.hass, self._context)
         await self._async_update_soon()
 
     async def async_turn_off(self) -> None:
